🤖 refactor: simplify workflow action helpers
Move built-in Git workflow actions into generated action source files, consolidate GitHub helper calls, avoid an extra label refetch, and reuse the shared plain-object guard for workflow args.
