+++ src/node/builtinWorkflowActions/github/getIssueAutomationState.js
@@ -4,7 +4,10 @@ export const metadata = {
   effect: "external",
   inputSchema: { type: "object" },
   outputSchema: { type: "object" },
-  permissions: [{ kind: "command", command: "gh api" }, { kind: "command", command: "gh issue view" }],
+  permissions: [
+    { kind: "command", command: "gh api" },
+    { kind: "command", command: "gh issue view" },
+  ],
   timeoutMs: 60000,
 };
 
@@ -17,9 +20,12 @@ export async function execute(rawInput, ctx) {
   const marker = requiredString(input.marker, "marker");
   const markerKey = requiredString(input.markerKey, "markerKey");
   const promptVersion = optionalString(input.promptVersion) || "v1";
-  const issue = await ctx.execJson("gh", ["issue", "view", String(number), "--repo", repository, "--json", "labels"]);
+  const [issue, comments] = await Promise.all([
+    getIssueView(ctx, repository, number, ["labels"]),
+    listComments(ctx, parts.owner, parts.repo, number),
+  ]);
   const labelNames = normalizeIssue(issue).labelNames;
-  const matching = (await listComments(ctx, parts.owner, parts.repo, number)).filter((comment) =>
+  const matching = comments.filter((comment) =>
     isMatchingMarker(comment.body, marker, markerKey, promptVersion)
   );
   const statuses = matching.map((comment) => markerStatus(comment.body)).filter(Boolean);
@@ -28,6 +34,10 @@ export async function execute(rawInput, ctx) {
     promptStarted: statuses.includes("prompt-started"),
     reportPosted: statuses.includes("report-posted"),
     labelNames,
-    markerComments: matching.map((comment) => ({ id: comment.id, url: comment.html_url || null, status: markerStatus(comment.body) })),
+    markerComments: matching.map((comment) => ({
+      id: comment.id,
+      url: comment.html_url || null,
+      status: markerStatus(comment.body),
+    })),
   };
 }
