+++ src/node/services/workflows/builtInWorkflowActionContent.generated.ts
@@ -3,16 +3,30 @@
 // Source: src/node/builtinWorkflowActions/**/*.js
 
 export const BUILTIN_WORKFLOW_ACTION_CONTENT: Record<string, string> = {
+  "git.changedFiles":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Return changed file lists for branch, staged, unstaged, and untracked Git state",\n  effect: "read",\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "git diff" },\n    { kind: "command", command: "git ls-files" },\n  ],\n  timeoutMs: 10000,\n};\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const head = optionalString(input.head) ?? "HEAD";\n  const staged = parseNameStatus(await runGit(ctx, ["diff", "--name-status", "--staged"]));\n  const unstaged = parseNameStatus(await runGit(ctx, ["diff", "--name-status"]));\n  const untrackedOutput = await runGit(ctx, ["ls-files", "--others", "--exclude-standard"]);\n  const untracked =\n    untrackedOutput.length === 0 ? [] : untrackedOutput.split(/\\r?\\n/).filter(Boolean);\n  const base = await tryResolveBase(ctx, input);\n  if (base == null) {\n    return { base: null, head, mergeBase: null, branch: [], staged, unstaged, untracked };\n  }\n  const mergeBase = await resolveMergeBase(ctx, base, head);\n  const branch = parseNameStatus(\n    await runGit(ctx, ["diff", "--name-status", mergeBase + ".." + head])\n  );\n  return { base, head, mergeBase, branch, staged, unstaged, untracked };\n};\n',
+  "git.commitsBetween":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Return commits reachable from head but not from the trunk/base branch",\n  effect: "read",\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "git log" }],\n  timeoutMs: 10000,\n};\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const base = await resolveBase(ctx, input);\n  const head = optionalString(input.head) ?? "HEAD";\n  const limit = boundedLimit(input.limit, 100);\n  const mergeBase = await resolveMergeBase(ctx, base, head);\n  const stdout = await runGit(ctx, [\n    "log",\n    "--max-count=" + String(limit),\n    "--format=%H%x1f%h%x1f%an%x1f%ae%x1f%aI%x1f%s%x1e",\n    mergeBase + ".." + head,\n  ]);\n  const commits = stdout\n    .split("\\x1e")\n    .map((record) => record.trim())\n    .filter((record) => record.length > 0)\n    .map((record) => {\n      const [hash, shortHash, authorName, authorEmail, authoredAt, subject] = record.split("\\x1f");\n      return { hash, shortHash, authorName, authorEmail, authoredAt, subject };\n    });\n  return { base, head, mergeBase, commits, count: commits.length };\n};\n',
+  "git.diff":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Return git diff output for branch, staged, and unstaged changes",\n  effect: "read",\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "git diff" }],\n  timeoutMs: 10000,\n};\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const head = optionalString(input.head) ?? "HEAD";\n  const staged = await captureGit(ctx, ["diff", "--staged"], [0]);\n  const unstaged = await captureGit(ctx, ["diff"], [0]);\n  const base = await tryResolveBase(ctx, input);\n  if (base == null) {\n    return {\n      base: null,\n      head,\n      mergeBase: null,\n      branch: "",\n      staged: staged.text,\n      unstaged: unstaged.text,\n      truncated: { branch: false, staged: staged.truncated, unstaged: unstaged.truncated },\n    };\n  }\n  const mergeBase = await resolveMergeBase(ctx, base, head);\n  const branch = await captureGit(ctx, ["diff", mergeBase + ".." + head], [0]);\n  return {\n    base,\n    head,\n    mergeBase,\n    branch: branch.text,\n    staged: staged.text,\n    unstaged: unstaged.text,\n    truncated: {\n      branch: branch.truncated,\n      staged: staged.truncated,\n      unstaged: unstaged.truncated,\n    },\n  };\n};\n',
+  "git.diffStat":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Return git diff --stat output for branch, staged, and unstaged changes",\n  effect: "read",\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "git diff" }],\n  timeoutMs: 10000,\n};\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const head = optionalString(input.head) ?? "HEAD";\n  const staged = await runGit(ctx, ["diff", "--stat", "--staged"]);\n  const unstaged = await runGit(ctx, ["diff", "--stat"]);\n  const base = await tryResolveBase(ctx, input);\n  if (base == null) {\n    return { base: null, head, mergeBase: null, branch: null, staged, unstaged };\n  }\n  const mergeBase = await resolveMergeBase(ctx, base, head);\n  const branch = await runGit(ctx, ["diff", "--stat", mergeBase + ".." + head]);\n  return { base, head, mergeBase, branch, staged, unstaged };\n};\n',
+  "git.preflight":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Validate that the current Git checkout is safe for workflow patch application",\n  effect: "read",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "git status" },\n    { kind: "command", command: "git rev-parse" },\n  ],\n  timeoutMs: 10000,\n};\n\nfunction normalizedBranch(branch) {\n  if (typeof branch !== "string") return "";\n  const trimmed = branch.trim();\n  return trimmed && trimmed !== "HEAD (no branch)" ? trimmed : "";\n}\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const status = await readStatus(ctx, input);\n  const expectedBranch = optionalString(input.expectedBranch);\n  const expectedHeadSha = optionalString(input.expectedHeadSha);\n  const requireClean = input.requireClean !== false && input.allowDirty !== true;\n  if (expectedBranch && normalizedBranch(status.branch) !== expectedBranch) {\n    return {\n      ok: false,\n      reason:\n        "Current branch " +\n        (status.branch || "unknown") +\n        " does not match expected branch " +\n        expectedBranch,\n      status,\n      expectedBranch,\n      expectedHeadSha,\n    };\n  }\n  if (expectedHeadSha && status.headSha !== expectedHeadSha) {\n    return {\n      ok: false,\n      reason:\n        "Current HEAD " +\n        (status.headSha || "unknown") +\n        " does not match expected HEAD " +\n        expectedHeadSha,\n      status,\n      expectedBranch,\n      expectedHeadSha,\n    };\n  }\n  if (requireClean && !status.clean) {\n    return {\n      ok: false,\n      reason: "Current worktree is dirty; commit or stash changes before applying workflow patches",\n      status,\n      expectedBranch,\n      expectedHeadSha,\n    };\n  }\n  return { ok: true, reason: "", status, expectedBranch, expectedHeadSha };\n};\n',
+  "git.reviewContext":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Return a compact review-ready Git context snapshot",\n  effect: "read",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "git status" },\n    { kind: "command", command: "git diff" },\n    { kind: "command", command: "git log" },\n    { kind: "command", command: "git ls-files" },\n  ],\n  timeoutMs: 10000,\n};\n\nconst DEFAULT_DIFF_CHAR_BUDGET = 60000;\nconst DEFAULT_METADATA_CHAR_BUDGET = 20000;\nconst DIFF_FIELDS = ["branch", "staged", "unstaged"];\n\nfunction boundedInt(value, fallback, min, max) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(min, Math.min(value, max));\n}\n\nasync function readGitReviewContext(ctx, input, diffBudget) {\n  const head = optionalString(input.head) ?? "HEAD";\n  const stagedFilesPromise = runGit(ctx, ["diff", "--name-status", "--staged"]).then(\n    parseNameStatus\n  );\n  const unstagedFilesPromise = runGit(ctx, ["diff", "--name-status"]).then(parseNameStatus);\n  const untrackedOutputPromise = runGit(ctx, ["ls-files", "--others", "--exclude-standard"]);\n  const base = await tryResolveBase(ctx, input);\n  const mergeBase = base == null ? null : await resolveMergeBase(ctx, base, head);\n  const [stagedFiles, unstagedFiles, untrackedOutput] = await Promise.all([\n    stagedFilesPromise,\n    unstagedFilesPromise,\n    untrackedOutputPromise,\n  ]);\n  const untracked =\n    untrackedOutput.length === 0 ? [] : untrackedOutput.split(/\\r?\\n/).filter(Boolean);\n  const shouldReadDiff = diffBudget > 0;\n  const branchFilesPromise =\n    mergeBase == null\n      ? Promise.resolve([])\n      : runGit(ctx, ["diff", "--name-status", mergeBase + ".." + head]).then(parseNameStatus);\n  const branchDiffPromise =\n    mergeBase == null\n      ? Promise.resolve({ text: "", truncated: false })\n      : shouldReadDiff\n        ? captureGit(ctx, ["diff", mergeBase + ".." + head], [0])\n        : branchFilesPromise.then((branchFiles) => ({\n            text: "",\n            truncated: branchFiles.length > 0,\n          }));\n  // Keep ctx.exec fan-out bounded; the action runner tracks each spawned child.\n  const [branchFiles, stagedDiff, unstagedDiff, branchDiff] = await Promise.all([\n    branchFilesPromise,\n    shouldReadDiff\n      ? captureGit(ctx, ["diff", "--staged"], [0])\n      : { text: "", truncated: stagedFiles.length > 0 },\n    shouldReadDiff\n      ? captureGit(ctx, ["diff"], [0])\n      : { text: "", truncated: unstagedFiles.length > 0 },\n    branchDiffPromise,\n  ]);\n  const [stagedStat, unstagedStat, branchStat, commits] = await Promise.all([\n    runGit(ctx, ["diff", "--stat", "--staged"]),\n    runGit(ctx, ["diff", "--stat"]),\n    mergeBase == null ? "" : runGit(ctx, ["diff", "--stat", mergeBase + ".." + head]),\n    input.includeCommits === true && mergeBase != null\n      ? readCommits(\n          ctx,\n          mergeBase,\n          head,\n          boundedInt(input.commitLimit ?? input.commitsLimit, 20, 1, 100)\n        )\n      : [],\n  ]);\n  return {\n    base,\n    head,\n    mergeBase,\n    changedFiles: {\n      base,\n      head,\n      mergeBase,\n      branch: branchFiles,\n      staged: stagedFiles,\n      unstaged: unstagedFiles,\n      untracked,\n    },\n    diffStat: {\n      base,\n      head,\n      mergeBase,\n      branch: branchStat,\n      staged: stagedStat,\n      unstaged: unstagedStat,\n    },\n    diff: {\n      base,\n      head,\n      mergeBase,\n      branch: branchDiff.text,\n      staged: stagedDiff.text,\n      unstaged: unstagedDiff.text,\n      truncated: {\n        branch: branchDiff.truncated,\n        staged: stagedDiff.truncated,\n        unstaged: unstagedDiff.truncated,\n      },\n    },\n    commits: { base, head, mergeBase, commits, count: commits.length },\n  };\n}\n\nasync function readCommits(ctx, mergeBase, head, limit) {\n  const stdout = await runGit(ctx, [\n    "log",\n    "--max-count=" + String(limit),\n    "--format=%H%x1f%h%x1f%an%x1f%ae%x1f%aI%x1f%s%x1e",\n    mergeBase + ".." + head,\n  ]);\n  return stdout\n    .split("\\x1e")\n    .map((record) => record.trim())\n    .filter((record) => record.length > 0)\n    .map((record) => {\n      const [hash, shortHash, authorName, authorEmail, authoredAt, subject] = record.split("\\x1f");\n      return { hash, shortHash, authorName, authorEmail, authoredAt, subject };\n    });\n}\n\nfunction allChangedFiles(changedFiles, status) {\n  const files = [];\n  addFileEntries(files, changedFiles.branch);\n  addFileEntries(files, changedFiles.staged);\n  addFileEntries(files, changedFiles.unstaged);\n  addFilePaths(files, changedFiles.untracked);\n  addFileEntries(files, status && status.staged);\n  addFileEntries(files, status && status.unstaged);\n  addFilePaths(files, status && status.untracked);\n  return files;\n}\n\nfunction addFileEntries(files, entries) {\n  if (!Array.isArray(entries)) return;\n  for (const entry of entries) {\n    if (entry && typeof entry === "object") {\n      addFilePath(files, entry.path);\n      addFilePath(files, entry.oldPath);\n    }\n  }\n}\n\nfunction addFilePaths(files, paths) {\n  if (!Array.isArray(paths)) return;\n  for (const path of paths) addFilePath(files, path);\n}\n\nfunction addFilePath(files, path) {\n  if (typeof path !== "string") return;\n  const trimmed = path.trim();\n  if (trimmed.length === 0 || files.includes(trimmed)) return;\n  files.push(trimmed);\n}\n\nfunction compactDiff(diff, budget) {\n  const compacted = {\n    base: diff.base,\n    head: diff.head,\n    mergeBase: diff.mergeBase,\n    truncated: diff.truncated,\n    workflowBudgetChars: budget,\n    workflowCompactions: [],\n  };\n  let remaining = budget;\n  for (const field of DIFF_FIELDS) {\n    const value = diff[field];\n    if (typeof value !== "string") {\n      compacted[field] = value;\n      continue;\n    }\n    const included = Math.max(0, Math.min(value.length, remaining));\n    compacted[field] =\n      included === value.length\n        ? value\n        : value.slice(0, included) +\n          "\\n\\n[Workflow prompt budget omitted the rest of the " +\n          field +\n          " diff.]";\n    remaining -= included;\n    if (included < value.length)\n      compacted.workflowCompactions.push({\n        field,\n        originalChars: value.length,\n        includedChars: included,\n      });\n  }\n  return compacted;\n}\n\nfunction compactText(value, limit) {\n  if (typeof value !== "string" || value.length <= limit) return value;\n  return (\n    value.slice(0, limit) +\n    "\\n\\n[Workflow metadata budget omitted " +\n    (value.length - limit) +\n    " chars.]"\n  );\n}\n\nfunction renderSnapshot(context, files, failures) {\n  const sections = [];\n  const status = context.status || {};\n  sections.push(\n    "Repository status: branch " +\n      (status.branch || "unknown") +\n      (status.upstream ? " tracking " + status.upstream : "") +\n      "; staged " +\n      arrayLength(status.staged) +\n      "; unstaged " +\n      arrayLength(status.unstaged) +\n      "; untracked " +\n      arrayLength(status.untracked)\n  );\n  if (files.length > 0) sections.push("Changed files: " + files.join(", "));\n  if (\n    context.commits &&\n    Array.isArray(context.commits.commits) &&\n    context.commits.commits.length > 0\n  ) {\n    sections.push(\n      "Commits since " +\n        (context.commits.base || "unknown") +\n        ":\\n" +\n        context.commits.commits\n          .map((commit) => "- " + (commit.shortHash || "unknown") + " " + (commit.subject || ""))\n          .join("\\n")\n    );\n  }\n  const statSections = [];\n  if (hasText(context.diffStat && context.diffStat.branch))\n    statSections.push("Branch diff stat:\\n" + context.diffStat.branch);\n  if (hasText(context.diffStat && context.diffStat.staged))\n    statSections.push("Staged diff stat:\\n" + context.diffStat.staged);\n  if (hasText(context.diffStat && context.diffStat.unstaged))\n    statSections.push("Unstaged diff stat:\\n" + context.diffStat.unstaged);\n  if (statSections.length > 0) sections.push(statSections.join("\\n\\n"));\n  if (arrayLength(status.untracked) > 0)\n    sections.push(\n      "Untracked file contents are not included in the automatic diff snapshot; only their paths are visible."\n    );\n  if (failures.length > 0)\n    sections.push(\n      "Git context warnings:\\n" +\n        failures.map((failure) => "- " + failure.action + ": " + failure.error).join("\\n")\n    );\n  return sections.join("\\n\\n");\n}\n\nfunction renderDiff(diff) {\n  const parts = [];\n  if (hasText(diff.branch))\n    parts.push(\n      "Branch diff (" +\n        (diff.base || "unknown") +\n        ".." +\n        (diff.head || "unknown") +\n        ")\\n" +\n        diff.branch\n    );\n  if (hasText(diff.staged)) parts.push("Staged diff\\n" + diff.staged);\n  if (hasText(diff.unstaged)) parts.push("Unstaged diff\\n" + diff.unstaged);\n  return parts.join("\\n\\n");\n}\n\nfunction hasText(value) {\n  return typeof value === "string" && value.trim().length > 0;\n}\nfunction arrayLength(value) {\n  return Array.isArray(value) ? value.length : 0;\n}\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const failures = [];\n  let status = null;\n  let context = null;\n  const diffBudget = boundedInt(input.diffCharBudget, DEFAULT_DIFF_CHAR_BUDGET, 0, 500000);\n  try {\n    status = await readStatus(ctx, input, { includeIgnored: input.includeIgnored === true });\n  } catch (error) {\n    failures.push({ action: "git.status", error: String((error && error.message) || error) });\n  }\n  try {\n    context = await readGitReviewContext(ctx, input, diffBudget);\n  } catch (error) {\n    failures.push({\n      action: "git.reviewContext",\n      error: String((error && error.message) || error),\n    });\n    context = {\n      base: null,\n      head: optionalString(input.head) ?? "HEAD",\n      mergeBase: null,\n      changedFiles: { branch: [], staged: [], unstaged: [], untracked: [] },\n      diffStat: { branch: "", staged: "", unstaged: "" },\n      diff: {\n        branch: "",\n        staged: "",\n        unstaged: "",\n        truncated: { branch: false, staged: false, unstaged: false },\n      },\n      commits: { commits: [], count: 0 },\n    };\n  }\n  const files = allChangedFiles(context.changedFiles, status);\n  context.changedFiles.all = files;\n  context.status = status;\n  context.failures = failures;\n  const compactedDiff = compactDiff(context.diff, diffBudget);\n  const flags = {\n    hasChanges:\n      files.length > 0 ||\n      hasText(context.diff.branch) ||\n      hasText(context.diff.staged) ||\n      hasText(context.diff.unstaged),\n    hasUncommittedChanges:\n      arrayLength(status && status.staged) > 0 ||\n      arrayLength(status && status.unstaged) > 0 ||\n      arrayLength(status && status.untracked) > 0,\n    hasUntrackedChanges:\n      arrayLength(status && status.untracked) > 0 ||\n      arrayLength(context.changedFiles.untracked) > 0,\n    hasOnlyUntrackedChanges:\n      files.length > 0 &&\n      arrayLength(context.changedFiles.branch) === 0 &&\n      arrayLength(context.changedFiles.staged) === 0 &&\n      arrayLength(context.changedFiles.unstaged) === 0 &&\n      !hasText(context.diff.branch) &&\n      !hasText(context.diff.staged) &&\n      !hasText(context.diff.unstaged),\n    clean: Boolean(status && status.clean),\n  };\n  const snapshotMarkdown = compactText(\n    renderSnapshot(context, files, failures),\n    boundedInt(input.metadataCharBudget, DEFAULT_METADATA_CHAR_BUDGET, 0, 500000)\n  );\n  return Object.assign({}, context, {\n    diff: compactedDiff,\n    flags,\n    rendered: {\n      snapshotMarkdown,\n      diffMarkdown: renderDiff(compactedDiff),\n      compactJson: JSON.stringify(\n        { status, changedFiles: context.changedFiles, diffStat: context.diffStat, failures, flags },\n        null,\n        2\n      ),\n    },\n    compactions: compactedDiff.workflowCompactions,\n  });\n};\n',
+  "git.status":
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.length > 0 ? value : undefined;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nasync function runGit(ctx, args) {\n  const result = await ctx.execChecked("git", args);\n  return result.stdout.trimEnd();\n}\n\nasync function tryGit(ctx, args) {\n  const result = await ctx.exec("git", args);\n  if (result.stdoutTruncated || result.stderrTruncated) {\n    throw new Error("git command output exceeded workflow action capture limit");\n  }\n  return result.exitCode === 0 ? result.stdout.trimEnd() : null;\n}\n\nasync function captureGit(ctx, args, allowedExitCodes) {\n  const result = await ctx.exec("git", args);\n  if (!allowedExitCodes.includes(result.exitCode)) {\n    throw new Error((result.stderr || result.stdout || "git command failed").trim());\n  }\n  return {\n    text: result.stdout.trimEnd(),\n    truncated: result.stdoutTruncated || result.stderrTruncated,\n  };\n}\n\nasync function resolveBase(ctx, input) {\n  const explicitBase = optionalString(input.base) ?? optionalString(input.trunk);\n  if (explicitBase != null) return explicitBase;\n  const originHead = await tryGit(ctx, [\n    "symbolic-ref",\n    "--quiet",\n    "--short",\n    "refs/remotes/origin/HEAD",\n  ]);\n  if (originHead != null && originHead.length > 0) return originHead;\n  for (const candidate of ["main", "master", "trunk"]) {\n    if ((await tryGit(ctx, ["rev-parse", "--verify", candidate])) != null) return candidate;\n  }\n  throw new Error("Unable to determine trunk branch; pass input.base or input.trunk");\n}\n\nasync function tryResolveBase(ctx, input) {\n  try {\n    return await resolveBase(ctx, input);\n  } catch {\n    return null;\n  }\n}\n\nasync function resolveMergeBase(ctx, base, head) {\n  return await runGit(ctx, ["merge-base", base, head]);\n}\n\nfunction parseNameStatus(stdout) {\n  if (stdout.length === 0) return [];\n  return stdout\n    .split(/\\r?\\n/)\n    .filter(Boolean)\n    .map((line) => {\n      const parts = line.split("\\t");\n      const status = parts[0] || "";\n      if (parts.length >= 3) {\n        return { status, oldPath: parts[1], path: parts[2] };\n      }\n      return { status, path: parts[1] || "" };\n    });\n}\n\nfunction parseBranchHeader(line) {\n  let branchText = line.slice(3);\n  let ahead = 0;\n  let behind = 0;\n  const trackingMatch = branchText.match(/ \\[(.+)\\]$/);\n  if (trackingMatch != null) {\n    branchText = branchText.slice(0, -trackingMatch[0].length);\n    for (const part of trackingMatch[1].split(", ")) {\n      const aheadMatch = part.match(/^ahead (\\d+)$/);\n      const behindMatch = part.match(/^behind (\\d+)$/);\n      if (aheadMatch != null) ahead = Number(aheadMatch[1]);\n      if (behindMatch != null) behind = Number(behindMatch[1]);\n    }\n  }\n  const [rawBranch, upstream] = branchText.split("...");\n  const branch = rawBranch.replace(/^No commits yet on /, "");\n  return { branch, upstream: upstream || null, ahead, behind };\n}\n\nfunction parseStatusLine(line) {\n  const index = line[0] || " ";\n  const worktree = line[1] || " ";\n  const rawPath = line.slice(3);\n  const renameParts = rawPath.split(" -> ");\n  return {\n    status: (index + worktree).trim(),\n    index,\n    worktree,\n    path: renameParts[renameParts.length - 1],\n    oldPath: renameParts.length > 1 ? renameParts[0] : undefined,\n  };\n}\n\nasync function readStatus(ctx, input, options) {\n  const includeIgnored = Boolean(options && options.includeIgnored);\n  const requestedHead = optionalString(input.head) ?? "HEAD";\n  const [headSha, requestedHeadSha, requestedHeadRef, stdout] = await Promise.all([\n    tryGit(ctx, ["rev-parse", "--verify", "HEAD"]),\n    tryGit(ctx, ["rev-parse", "--verify", requestedHead]),\n    tryGit(ctx, ["rev-parse", "--symbolic-full-name", "--verify", requestedHead]),\n    runGit(ctx, [\n      "status",\n      "--porcelain=v1",\n      "-b",\n      "-uall",\n      includeIgnored ? "--ignored=traditional" : "--ignored=no",\n      "--ahead-behind",\n    ]),\n  ]);\n  const lines = stdout.split(/\\r?\\n/).filter(Boolean);\n  const header = lines[0]?.startsWith("## ")\n    ? parseBranchHeader(lines[0])\n    : { branch: null, upstream: null, ahead: 0, behind: 0 };\n  const staged = [];\n  const unstaged = [];\n  const untracked = [];\n  const ignored = [];\n  for (const line of lines.slice(\n    header.branch == null && lines[0]?.startsWith("## ") !== true ? 0 : 1\n  )) {\n    const file = parseStatusLine(line);\n    if (file.index === "?" && file.worktree === "?") {\n      untracked.push(file.path);\n      continue;\n    }\n    if (file.index === "!" && file.worktree === "!") {\n      ignored.push(file.path);\n      continue;\n    }\n    if (file.index !== " " && file.index !== "?") staged.push(file);\n    if (file.worktree !== " " && file.worktree !== "?") unstaged.push(file);\n  }\n  return {\n    ...header,\n    headSha,\n    requestedHead,\n    requestedHeadSha,\n    requestedHeadRef,\n    clean: staged.length === 0 && unstaged.length === 0 && untracked.length === 0,\n    staged,\n    unstaged,\n    untracked,\n    ignored,\n  };\n}\n\n\nmodule.exports.metadata = {\n  version: 1,\n  description: "Return branch, upstream, and working tree status for the current Git repository",\n  effect: "read",\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "git status" },\n    { kind: "command", command: "git rev-parse" },\n  ],\n  timeoutMs: 10000,\n};\n\nmodule.exports.execute = async function (rawInput, ctx) {\n  const input = inputObject(rawInput);\n  return await readStatus(ctx, input, { includeIgnored: input.includeIgnored !== false });\n};\n',
   "github.ensureIssueLabels":
-    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Idempotently add and remove GitHub issue labels",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "gh issue edit" },\n    { kind: "command", command: "gh issue view" },\n  ],\n  timeoutMs: 60000,\n};\n\nasync function getLabelNames(ctx, repository, number) {\n  const args = ["issue", "view", String(number), "--json", "labels"];\n  if (repository) args.push("--repo", repository);\n  const issue = await ctx.execJson("gh", args);\n  return normalizeIssue(issue).labelNames;\n}\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = repositoryFromInput(input);\n  const number = requiredIssueNumber(input.number);\n  const addLabels = stringList(input.addLabels);\n  const removeLabels = stringList(input.removeLabels);\n  const before = await getLabelNames(ctx, repository, number);\n  const missingAddLabels = addLabels.filter((label) => !before.includes(label));\n  const presentRemoveLabels = removeLabels.filter((label) => before.includes(label));\n  if (missingAddLabels.length === 0 && presentRemoveLabels.length === 0) {\n    return { changed: false, before, after: before, added: [], removed: [] };\n  }\n  const args = ["issue", "edit", String(number)];\n  if (repository) args.push("--repo", repository);\n  for (const label of missingAddLabels) args.push("--add-label", label);\n  for (const label of presentRemoveLabels) args.push("--remove-label", label);\n  await ctx.execChecked("gh", args);\n  const after = await getLabelNames(ctx, repository, number);\n  return {\n    changed: true,\n    before,\n    after,\n    added: missingAddLabels,\n    removed: presentRemoveLabels,\n  };\n}\n\nexport const reconcile = execute;\n',
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function getIssueView(ctx, repository, number, fields) {\n  const args = ["issue", "view", String(number), "--json", fields.join(",")];\n  if (repository) args.push("--repo", repository);\n  return await ctx.execJson("gh", args);\n}\n\nasync function fetchCommentsPage(ctx, owner, repo, number, page) {\n  return await ctx.execJson("gh", [\n    "api",\n    "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n  ]);\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Idempotently add and remove GitHub issue labels",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "gh issue edit" },\n    { kind: "command", command: "gh issue view" },\n  ],\n  timeoutMs: 60000,\n};\n\nasync function getLabelNames(ctx, repository, number) {\n  const issue = await getIssueView(ctx, repository, number, ["labels"]);\n  return normalizeIssue(issue).labelNames;\n}\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = repositoryFromInput(input);\n  const number = requiredIssueNumber(input.number);\n  const addLabels = stringList(input.addLabels);\n  const removeLabels = stringList(input.removeLabels);\n  const before = await getLabelNames(ctx, repository, number);\n  const missingAddLabels = addLabels.filter((label) => !before.includes(label));\n  const presentRemoveLabels = removeLabels.filter((label) => before.includes(label));\n  if (missingAddLabels.length === 0 && presentRemoveLabels.length === 0) {\n    return { changed: false, before, after: before, added: [], removed: [] };\n  }\n  const args = ["issue", "edit", String(number)];\n  if (repository) args.push("--repo", repository);\n  for (const label of missingAddLabels) args.push("--add-label", label);\n  for (const label of presentRemoveLabels) args.push("--remove-label", label);\n  await ctx.execChecked("gh", args);\n  const after = before.filter((label) => !presentRemoveLabels.includes(label));\n  for (const label of missingAddLabels) {\n    if (!after.includes(label)) after.push(label);\n  }\n  return {\n    changed: true,\n    before,\n    after,\n    added: missingAddLabels,\n    removed: presentRemoveLabels,\n  };\n}\n\nexport const reconcile = execute;\n',
   "github.getIssueAutomationState":
-    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Read GitHub issue automation marker comments and done labels",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh api" }, { kind: "command", command: "gh issue view" }],\n  timeoutMs: 60000,\n};\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = requiredRepository(input);\n  const parts = splitRepository(repository);\n  const number = requiredIssueNumber(input.number);\n  const doneLabels = stringList(input.doneLabels);\n  const marker = requiredString(input.marker, "marker");\n  const markerKey = requiredString(input.markerKey, "markerKey");\n  const promptVersion = optionalString(input.promptVersion) || "v1";\n  const issue = await ctx.execJson("gh", ["issue", "view", String(number), "--repo", repository, "--json", "labels"]);\n  const labelNames = normalizeIssue(issue).labelNames;\n  const matching = (await listComments(ctx, parts.owner, parts.repo, number)).filter((comment) =>\n    isMatchingMarker(comment.body, marker, markerKey, promptVersion)\n  );\n  const statuses = matching.map((comment) => markerStatus(comment.body)).filter(Boolean);\n  return {\n    done: doneLabels.some((label) => labelNames.includes(label)),\n    promptStarted: statuses.includes("prompt-started"),\n    reportPosted: statuses.includes("report-posted"),\n    labelNames,\n    markerComments: matching.map((comment) => ({ id: comment.id, url: comment.html_url || null, status: markerStatus(comment.body) })),\n  };\n}\n',
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function getIssueView(ctx, repository, number, fields) {\n  const args = ["issue", "view", String(number), "--json", fields.join(",")];\n  if (repository) args.push("--repo", repository);\n  return await ctx.execJson("gh", args);\n}\n\nasync function fetchCommentsPage(ctx, owner, repo, number, page) {\n  return await ctx.execJson("gh", [\n    "api",\n    "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n  ]);\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Read GitHub issue automation marker comments and done labels",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "gh api" },\n    { kind: "command", command: "gh issue view" },\n  ],\n  timeoutMs: 60000,\n};\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = requiredRepository(input);\n  const parts = splitRepository(repository);\n  const number = requiredIssueNumber(input.number);\n  const doneLabels = stringList(input.doneLabels);\n  const marker = requiredString(input.marker, "marker");\n  const markerKey = requiredString(input.markerKey, "markerKey");\n  const promptVersion = optionalString(input.promptVersion) || "v1";\n  const [issue, comments] = await Promise.all([\n    getIssueView(ctx, repository, number, ["labels"]),\n    listComments(ctx, parts.owner, parts.repo, number),\n  ]);\n  const labelNames = normalizeIssue(issue).labelNames;\n  const matching = comments.filter((comment) =>\n    isMatchingMarker(comment.body, marker, markerKey, promptVersion)\n  );\n  const statuses = matching.map((comment) => markerStatus(comment.body)).filter(Boolean);\n  return {\n    done: doneLabels.some((label) => labelNames.includes(label)),\n    promptStarted: statuses.includes("prompt-started"),\n    reportPosted: statuses.includes("report-posted"),\n    labelNames,\n    markerComments: matching.map((comment) => ({\n      id: comment.id,\n      url: comment.html_url || null,\n      status: markerStatus(comment.body),\n    })),\n  };\n}\n',
   "github.getIssueConversation":
-    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Read a GitHub issue body and comments as markdown",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "gh issue view" },\n    { kind: "command", command: "gh api" },\n  ],\n  timeoutMs: 60000,\n};\n\nfunction repositoryPartsForComments(repository, issue) {\n  if (repository) return splitRepository(repository);\n  const match =\n    typeof issue.url === "string"\n      ? issue.url.match(/^https:\\/\\/github\\.com\\/([^/]+)\\/([^/]+)\\/issues\\/\\d+$/)\n      : null;\n  if (!match) throw new Error("repository or a GitHub issue URL is required to read comments");\n  return { owner: match[1], repo: match[2] };\n}\n\nfunction formatConversation(comments, commentBodyBudget, hasOmittedComments) {\n  const visibleComments = Array.isArray(comments) ? comments : [];\n  if (visibleComments.length === 0) return "(no issue comments)";\n  const markdown = visibleComments\n    .map(\n      (comment) =>\n        "### Comment by " +\n        ((comment.author && comment.author.login) || "unknown") +\n        "\\n\\n" +\n        truncateText(comment.body || "", commentBodyBudget)\n    )\n    .join("\\n\\n---\\n\\n");\n  return hasOmittedComments ? markdown + "\\n\\n[omitted additional comments]" : markdown;\n}\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = repositoryFromInput(input);\n  const number = requiredIssueNumber(input.number);\n  const maxComments = boundedLimit(input.maxComments, 100);\n  const issueBodyBudget = boundedCharBudget(\n    input.issueBodyCharBudget ?? input.bodyCharBudget,\n    10000\n  );\n  const commentBodyBudget = boundedCharBudget(input.commentBodyCharBudget, 10000);\n  const args = [\n    "issue",\n    "view",\n    String(number),\n    "--json",\n    "number,title,url,state,body,author,labels",\n  ];\n  if (repository) args.push("--repo", repository);\n  const issue = await ctx.execJson("gh", args);\n  const parts = repositoryPartsForComments(repository, issue);\n  const comments = await listComments(ctx, parts.owner, parts.repo, number, {\n    limit: maxComments + 1,\n  });\n  const visibleComments = comments.slice(0, maxComments);\n  const hasOmittedComments = comments.length > visibleComments.length;\n  const normalizedIssue = normalizeIssue(issue);\n  return {\n    repository: repository || null,\n    number,\n    issue: { ...normalizedIssue, body: truncateText(normalizedIssue.body, issueBodyBudget) },\n    conversationMarkdown: formatConversation(\n      visibleComments,\n      commentBodyBudget,\n      hasOmittedComments\n    ),\n    limits: { maxComments, issueBodyBudget, commentBodyBudget, hasOmittedComments },\n  };\n}\n',
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function getIssueView(ctx, repository, number, fields) {\n  const args = ["issue", "view", String(number), "--json", fields.join(",")];\n  if (repository) args.push("--repo", repository);\n  return await ctx.execJson("gh", args);\n}\n\nasync function fetchCommentsPage(ctx, owner, repo, number, page) {\n  return await ctx.execJson("gh", [\n    "api",\n    "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n  ]);\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Read a GitHub issue body and comments as markdown",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [\n    { kind: "command", command: "gh issue view" },\n    { kind: "command", command: "gh api" },\n  ],\n  timeoutMs: 60000,\n};\n\nfunction repositoryPartsForComments(repository, issue) {\n  if (repository) return splitRepository(repository);\n  const match =\n    typeof issue.url === "string"\n      ? issue.url.match(/^https:\\/\\/github\\.com\\/([^/]+)\\/([^/]+)\\/issues\\/\\d+$/)\n      : null;\n  if (!match) throw new Error("repository or a GitHub issue URL is required to read comments");\n  return { owner: match[1], repo: match[2] };\n}\n\nfunction formatConversation(comments, commentBodyBudget, hasOmittedComments) {\n  const visibleComments = Array.isArray(comments) ? comments : [];\n  if (visibleComments.length === 0) return "(no issue comments)";\n  const markdown = visibleComments\n    .map(\n      (comment) =>\n        "### Comment by " +\n        ((comment.author && comment.author.login) || "unknown") +\n        "\\n\\n" +\n        truncateText(comment.body || "", commentBodyBudget)\n    )\n    .join("\\n\\n---\\n\\n");\n  return hasOmittedComments ? markdown + "\\n\\n[omitted additional comments]" : markdown;\n}\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = repositoryFromInput(input);\n  const number = requiredIssueNumber(input.number);\n  const maxComments = boundedLimit(input.maxComments, 100);\n  const issueBodyBudget = boundedCharBudget(\n    input.issueBodyCharBudget ?? input.bodyCharBudget,\n    10000\n  );\n  const commentBodyBudget = boundedCharBudget(input.commentBodyCharBudget, 10000);\n  const issueFields = ["number", "title", "url", "state", "body", "author", "labels"];\n  let issue;\n  let comments;\n  if (repository) {\n    const parts = splitRepository(repository);\n    [issue, comments] = await Promise.all([\n      getIssueView(ctx, repository, number, issueFields),\n      listComments(ctx, parts.owner, parts.repo, number, { limit: maxComments + 1 }),\n    ]);\n  } else {\n    issue = await getIssueView(ctx, repository, number, issueFields);\n    const parts = repositoryPartsForComments(repository, issue);\n    comments = await listComments(ctx, parts.owner, parts.repo, number, {\n      limit: maxComments + 1,\n    });\n  }\n  const visibleComments = comments.slice(0, maxComments);\n  const hasOmittedComments = comments.length > visibleComments.length;\n  const normalizedIssue = normalizeIssue(issue);\n  return {\n    repository: repository || null,\n    number,\n    issue: { ...normalizedIssue, body: truncateText(normalizedIssue.body, issueBodyBudget) },\n    conversationMarkdown: formatConversation(\n      visibleComments,\n      commentBodyBudget,\n      hasOmittedComments\n    ),\n    limits: { maxComments, issueBodyBudget, commentBodyBudget, hasOmittedComments },\n  };\n}\n',
   "github.listIssues":
-    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "List GitHub issues with reusable label/state filters",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh issue list" }],\n  timeoutMs: 60000,\n};\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = repositoryFromInput(input);\n  const state = optionalString(input.state) || "open";\n  const includeLabels = stringList(input.includeLabels);\n  const excludeLabels = stringList(input.excludeLabels);\n  const limit = boundedLimit(input.limit, 1000);\n  const includeBody = input.includeBody === true;\n  const bodyCharBudget = boundedCharBudget(input.bodyCharBudget, 2000);\n  const jsonFields =\n    "number,title,url,state,labels,author,createdAt,updatedAt" + (includeBody ? ",body" : "");\n  const args = ["issue", "list", "--state", state, "--limit", String(limit), "--json", jsonFields];\n  if (repository) args.push("--repo", repository);\n  for (const label of includeLabels) args.push("--label", label);\n  const issues = (await ctx.execJson("gh", args))\n    .map(normalizeIssue)\n    .map((issue) => ({\n      ...issue,\n      body: includeBody ? truncateText(issue.body, bodyCharBudget) : "",\n    }))\n    .filter((issue) => includeLabels.every((label) => issue.labelNames.includes(label)))\n    .filter((issue) => excludeLabels.every((label) => !issue.labelNames.includes(label)))\n    .sort((a, b) => a.number - b.number);\n  return {\n    repository: repository || null,\n    filters: { state, includeLabels, excludeLabels, limit, includeBody, bodyCharBudget },\n    issues,\n  };\n}\n',
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function getIssueView(ctx, repository, number, fields) {\n  const args = ["issue", "view", String(number), "--json", fields.join(",")];\n  if (repository) args.push("--repo", repository);\n  return await ctx.execJson("gh", args);\n}\n\nasync function fetchCommentsPage(ctx, owner, repo, number, page) {\n  return await ctx.execJson("gh", [\n    "api",\n    "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n  ]);\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "List GitHub issues with reusable label/state filters",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh issue list" }],\n  timeoutMs: 60000,\n};\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = repositoryFromInput(input);\n  const state = optionalString(input.state) || "open";\n  const includeLabels = stringList(input.includeLabels);\n  const excludeLabels = stringList(input.excludeLabels);\n  const limit = boundedLimit(input.limit, 1000);\n  const includeBody = input.includeBody === true;\n  const bodyCharBudget = boundedCharBudget(input.bodyCharBudget, 2000);\n  const jsonFields =\n    "number,title,url,state,labels,author,createdAt,updatedAt" + (includeBody ? ",body" : "");\n  const args = ["issue", "list", "--state", state, "--limit", String(limit), "--json", jsonFields];\n  if (repository) args.push("--repo", repository);\n  for (const label of includeLabels) args.push("--label", label);\n  const issues = (await ctx.execJson("gh", args))\n    .map(normalizeIssue)\n    .map((issue) => ({\n      ...issue,\n      body: includeBody ? truncateText(issue.body, bodyCharBudget) : "",\n    }))\n    .filter((issue) => includeLabels.every((label) => issue.labelNames.includes(label)))\n    .filter((issue) => excludeLabels.every((label) => !issue.labelNames.includes(label)))\n    .sort((a, b) => a.number - b.number);\n  return {\n    repository: repository || null,\n    filters: { state, includeLabels, excludeLabels, limit, includeBody, bodyCharBudget },\n    issues,\n  };\n}\n',
   "github.upsertIssueComment":
-    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Create or update a GitHub issue comment selected by marker",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh api" }],\n  timeoutMs: 60000,\n};\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = requiredRepository(input);\n  const parts = splitRepository(repository);\n  const number = requiredIssueNumber(input.number);\n  const marker = requiredString(input.marker, "marker");\n  const body = requiredString(input.body, "body");\n  const existing = await findComment(\n    ctx,\n    parts.owner,\n    parts.repo,\n    number,\n    (comment) => typeof comment.body === "string" && comment.body.includes(marker)\n  );\n  const payload = await ctx.writeTempJson({ body });\n  if (existing) {\n    await ctx.execChecked("gh", [\n      "api",\n      "-X",\n      "PATCH",\n      "repos/" + parts.owner + "/" + parts.repo + "/issues/comments/" + existing.id,\n      "--input",\n      payload.path,\n    ]);\n    return { action: "updated", commentId: existing.id, url: existing.html_url || null };\n  }\n  const created = await ctx.execJson("gh", [\n    "api",\n    "-X",\n    "POST",\n    "repos/" + parts.owner + "/" + parts.repo + "/issues/" + number + "/comments",\n    "--input",\n    payload.path,\n  ]);\n  return { action: "created", commentId: created.id || null, url: created.html_url || null };\n}\n\nexport const reconcile = execute;\n',
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function getIssueView(ctx, repository, number, fields) {\n  const args = ["issue", "view", String(number), "--json", fields.join(",")];\n  if (repository) args.push("--repo", repository);\n  return await ctx.execJson("gh", args);\n}\n\nasync function fetchCommentsPage(ctx, owner, repo, number, page) {\n  return await ctx.execJson("gh", [\n    "api",\n    "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n  ]);\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Create or update a GitHub issue comment selected by marker",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh api" }],\n  timeoutMs: 60000,\n};\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = requiredRepository(input);\n  const parts = splitRepository(repository);\n  const number = requiredIssueNumber(input.number);\n  const marker = requiredString(input.marker, "marker");\n  const body = requiredString(input.body, "body");\n  const existing = await findComment(\n    ctx,\n    parts.owner,\n    parts.repo,\n    number,\n    (comment) => typeof comment.body === "string" && comment.body.includes(marker)\n  );\n  const payload = await ctx.writeTempJson({ body });\n  if (existing) {\n    await ctx.execChecked("gh", [\n      "api",\n      "-X",\n      "PATCH",\n      "repos/" + parts.owner + "/" + parts.repo + "/issues/comments/" + existing.id,\n      "--input",\n      payload.path,\n    ]);\n    return { action: "updated", commentId: existing.id, url: existing.html_url || null };\n  }\n  const created = await ctx.execJson("gh", [\n    "api",\n    "-X",\n    "POST",\n    "repos/" + parts.owner + "/" + parts.repo + "/issues/" + number + "/comments",\n    "--input",\n    payload.path,\n  ]);\n  return { action: "created", commentId: created.id || null, url: created.html_url || null };\n}\n\nexport const reconcile = execute;\n',
   "github.verifyIssueCommentUrl":
-    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await ctx.execJson("gh", [\n      "api",\n      "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n    ]);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Verify that a GitHub issue comment URL belongs to an issue and contains expected text",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh api" }],\n  timeoutMs: 60000,\n};\n\nfunction parseCommentUrl(url) {\n  const match = String(url || "").match(/^https:\\/\\/github\\.com\\/([^/]+)\\/([^/]+)\\/issues\\/(\\d+)#issuecomment-(\\d+)$/);\n  return match ? { owner: match[1], repo: match[2], number: Number(match[3]), commentId: match[4] } : null;\n}\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = requiredRepository(input);\n  const parts = splitRepository(repository);\n  const number = requiredIssueNumber(input.number);\n  const parsed = parseCommentUrl(requiredString(input.url, "url"));\n  if (!parsed || parsed.owner !== parts.owner || parsed.repo !== parts.repo || parsed.number !== number) {\n    return { verified: false, reason: "comment-url-does-not-match-issue" };\n  }\n  const comment = await ctx.execJson("gh", ["api", "repos/" + parts.owner + "/" + parts.repo + "/issues/comments/" + parsed.commentId]);\n  const includes = stringList(input.requiredBodyIncludes);\n  const missing = includes.filter((text) => typeof comment.body !== "string" || !comment.body.includes(text));\n  if (missing.length > 0) return { verified: false, reason: "missing-required-body-text", missing, url: comment.html_url || input.url };\n  return { verified: true, reason: "", url: comment.html_url || input.url, commentId: comment.id || parsed.commentId };\n}\n',
+    'function inputObject(input) {\n  return input != null && typeof input === "object" && !Array.isArray(input) ? input : {};\n}\n\nfunction optionalString(value) {\n  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;\n}\n\nfunction requiredString(value, name) {\n  const text = optionalString(value);\n  if (!text) throw new Error(name + " must be a non-empty string");\n  return text;\n}\n\nfunction stringList(value) {\n  if (!Array.isArray(value)) return [];\n  return value\n    .filter((item) => typeof item === "string" && item.trim().length > 0)\n    .map((item) => item.trim());\n}\n\nfunction repositoryFromInput(input) {\n  const repository = optionalString(input.repository);\n  if (repository) return repository;\n  const owner = optionalString(input.owner);\n  const repo = optionalString(input.repo);\n  return owner && repo ? owner + "/" + repo : undefined;\n}\n\nfunction requiredRepository(input) {\n  const repository = repositoryFromInput(input);\n  if (!repository) throw new Error("repository or owner/repo is required");\n  return repository;\n}\n\nfunction splitRepository(repository) {\n  const parts = repository.split("/");\n  if (parts.length !== 2 || !parts[0] || !parts[1]) {\n    throw new Error("repository must use owner/repo format");\n  }\n  return { owner: parts[0], repo: parts[1] };\n}\n\nfunction requiredIssueNumber(value) {\n  if (!Number.isInteger(value) || value <= 0) {\n    throw new Error("number must be a positive integer issue number");\n  }\n  return value;\n}\n\nfunction boundedLimit(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(1, Math.min(value, 1000));\n}\n\nfunction boundedCharBudget(value, fallback) {\n  if (!Number.isInteger(value)) return fallback;\n  return Math.max(0, Math.min(value, 100000));\n}\n\nfunction truncateText(value, budget) {\n  const text = typeof value === "string" ? value : "";\n  if (text.length <= budget) return text;\n  return text.slice(0, budget) + "\\n\\n[truncated " + (text.length - budget) + " chars]";\n}\n\nfunction normalizeIssue(issue) {\n  const labelNames = Array.isArray(issue.labels)\n    ? issue.labels\n        .map((label) => (typeof label === "string" ? label : label.name))\n        .filter((name) => typeof name === "string" && name.length > 0)\n    : [];\n  return {\n    number: issue.number,\n    safeId: "issue-" + issue.number,\n    title: issue.title || "",\n    url: issue.url || "",\n    state: issue.state || "",\n    body: issue.body || "",\n    author: issue.author && issue.author.login ? issue.author.login : null,\n    createdAt: issue.createdAt || null,\n    updatedAt: issue.updatedAt || null,\n    labelNames,\n  };\n}\n\nfunction markerCommentNeedle(marker, markerKey, promptVersion) {\n  return "<!-- " + marker + " key=" + markerKey + " promptVersion=" + promptVersion;\n}\n\nfunction isMatchingMarker(body, marker, markerKey, promptVersion) {\n  return (\n    typeof body === "string" && body.includes(markerCommentNeedle(marker, markerKey, promptVersion))\n  );\n}\n\nfunction markerStatus(body) {\n  const match = typeof body === "string" ? body.match(/status=([a-z0-9_-]+)/i) : null;\n  return match ? match[1] : "";\n}\n\nasync function getIssueView(ctx, repository, number, fields) {\n  const args = ["issue", "view", String(number), "--json", fields.join(",")];\n  if (repository) args.push("--repo", repository);\n  return await ctx.execJson("gh", args);\n}\n\nasync function fetchCommentsPage(ctx, owner, repo, number, page) {\n  return await ctx.execJson("gh", [\n    "api",\n    "repos/" + owner + "/" + repo + "/issues/" + number + "/comments?per_page=100&page=" + page,\n  ]);\n}\n\nasync function listComments(ctx, owner, repo, number, options) {\n  const comments = [];\n  const limit = boundedLimit(options && options.limit, 1000);\n  for (let page = 1; page <= Math.ceil(limit / 100); page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    for (const comment of pageComments) {\n      comments.push(comment);\n      if (comments.length >= limit) return comments;\n    }\n    if (pageComments.length < 100) break;\n  }\n  return comments;\n}\n\nasync function findComment(ctx, owner, repo, number, predicate) {\n  for (let page = 1; page <= 10; page += 1) {\n    const pageComments = await fetchCommentsPage(ctx, owner, repo, number, page);\n    const match = pageComments.find(predicate);\n    if (match) return match;\n    if (pageComments.length < 100) break;\n  }\n  return undefined;\n}\n\n\nexport const metadata = {\n  version: 1,\n  description: "Verify that a GitHub issue comment URL belongs to an issue and contains expected text",\n  effect: "external",\n  inputSchema: { type: "object" },\n  outputSchema: { type: "object" },\n  permissions: [{ kind: "command", command: "gh api" }],\n  timeoutMs: 60000,\n};\n\nfunction parseCommentUrl(url) {\n  const match = String(url || "").match(/^https:\\/\\/github\\.com\\/([^/]+)\\/([^/]+)\\/issues\\/(\\d+)#issuecomment-(\\d+)$/);\n  return match ? { owner: match[1], repo: match[2], number: Number(match[3]), commentId: match[4] } : null;\n}\n\nexport async function execute(rawInput, ctx) {\n  const input = inputObject(rawInput);\n  const repository = requiredRepository(input);\n  const parts = splitRepository(repository);\n  const number = requiredIssueNumber(input.number);\n  const parsed = parseCommentUrl(requiredString(input.url, "url"));\n  if (!parsed || parsed.owner !== parts.owner || parsed.repo !== parts.repo || parsed.number !== number) {\n    return { verified: false, reason: "comment-url-does-not-match-issue" };\n  }\n  const comment = await ctx.execJson("gh", ["api", "repos/" + parts.owner + "/" + parts.repo + "/issues/comments/" + parsed.commentId]);\n  const includes = stringList(input.requiredBodyIncludes);\n  const missing = includes.filter((text) => typeof comment.body !== "string" || !comment.body.includes(text));\n  if (missing.length > 0) return { verified: false, reason: "missing-required-body-text", missing, url: comment.html_url || input.url };\n  return { verified: true, reason: "", url: comment.html_url || input.url, commentId: comment.id || parsed.commentId };\n}\n',
 };
